Some Buttondown price code moved to utils file.
